episode 24. Authorization With Policies

+++ app/Http/Controllers/ThreadsController.php
@@ -119,6 +119,7 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread); // 인증
         $thread->delete();
         
         if (request()->wantsJson()) {
